pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "cram"
version = "0.1.0"
description = "Canvas to Notion sync tool for university courses"
requires-python = ">=3.12"
dependencies = [
    "canvasapi>=3.3.0",
    "notion-client>=2.3.0",
    "markdownify>=0.13.1",
    "typer>=0.15.1",
    "python-dotenv>=1.1.0",
    "apscheduler>=3.11.0",
    "rich>=13.9.4",
]

[project.scripts]
cram = "cram.cli:app"

[tool.hatch.build.targets.wheel]
packages = ["cram"]
